+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2026-06-09
+
+### Fixed
+
+- Fixed issues with Sable sub-levels.
+- Backpacks can now be placed even when the offhand has an item.
+
 ## [1.7.0] - 2026-05-27
 
 ### Added